fix a buncha stuff
