+++ workspacesio/artifacts/crud.py
@@ -3,7 +3,6 @@ from typing import Optional
 from sqlalchemy import and_, func
 from sqlalchemy.orm import Session
 
-from workspacesio import schemas
+from workspacesio.common import schemas
 
 from . import models as artifact_models
-from . import schemas as artifact_schemas
