+++ workspacesio/cli/root.py
@@ -1,8 +1,7 @@
 import click
 from click_aliases import ClickAliasedGroup
 
-from workspacesio import schemas
-from workspacesio.indexing.producers import minio_list_root_children
+from workspacesio.common import schemas
 
 from .util import exit_with, handle_request_error
 
@@ -53,11 +52,16 @@ def make(cli: click.Group):
     @click.option("--index-all", is_flag=True)
     @click.pass_obj
     def import_all_workspaces(ctx, root_id, index_all):
+        # Dynamic, expensive imports
+        from workspacesio.common import producers
+
         r = ctx["session"].post("root/import", json={"root_id": root_id})
         if not r.ok:
             exit_with(handle_request_error(r))
         rdata = schemas.RootImport(**r.json())
-        root_contents = minio_list_root_children(node=rdata.node, root=rdata.root)
+        root_contents = producers.minio_list_root_children(
+            node=rdata.node, root=rdata.root
+        )
         workspace_list: List[schemas.WorkspaceDB] = []
         for folder in root_contents:
             prefix = folder.object_name.lstrip(rdata.root.base_path).strip("/")
