+++ workspacesio/common/__init__.py
@@ -1,0 +1,3 @@
+"""
+workspacesio.common library for CLI and server
+"""
