+++ workspacesio/database.py
@@ -5,7 +5,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
-from .schemas import UserDB
+from workspacesio.common.schemas import UserDB
 from .settings import settings
 
 database = databases.Database(settings.database_uri)
