+++ workspacesio/indexing/__init__.py
@@ -2,4 +2,4 @@
 Package for managing and handling index events and bucket notifications,
 and talking to elasticsearch
 """
-from . import api, models, producers, schemas, video
+from . import api, models
